+++ cms/src/config/router.config.ts
@@ -1,12 +1,9 @@
 import { createRouter } from "@tanstack/react-router";
 import { routeTree } from "../routeTree.gen";
 
-import App from "../app/App"; // Testing...
-
 const router = createRouter({
   routeTree,
   scrollRestoration: true,
-  Wrap: App,
   context: { auth: undefined! },
 });