+++ cms/src/routes/__root.tsx
@@ -1,13 +1,9 @@
-import {
-  createRootRoute,
-  createRootRouteWithContext,
-  Outlet,
-} from "@tanstack/react-router";
+import { createRootRouteWithContext } from "@tanstack/react-router";
 import DefaultLayout from "../layouts/DefaultLayout";
 import { AuthContext } from "../hooks/useAuth";
 
 interface RouterContext {
-  auth: AuthContext;
+  auth: AuthContext | null;
 }
 
 export const Route = createRootRouteWithContext<RouterContext>()({